- Better `Sync` boot up - Fixed a bug related to X86 CPU and atomic.AddUInt64 - Update config simple

+++ trap/core/sync/communication/messager/stats.go
@@ -22,6 +22,6 @@
 package messager
 
 type Stats struct {
-	TX int64
-	RX int64
+	TX int32
+	RX int32
 }
